requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
drf-yasg
django-filter
psycopg2-binary
python-dotenv
django-cors-headers
google-auth
requests
whitenoise